# services/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GleanWise",
    "description": "Local-first answer engine. `POST /query` then stream `GET /stream/{id}` (SSE).\nOpenAI-compatible endpoint: `POST /v1/chat/completions`.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/setup/test": {
      "post": {
        "tags": [
          "health"
        ],
        "summary": "Test settings before saving",
        "description": "Used by the first-run wizard: proves the model, key and search backend actually work.",
        "operationId": "setup_test_setup_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupTestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Start a question",
        "description": "Returns immediately. Open `stream_url` (Server-Sent Events) to receive the answer. Re-connect with `Last-Event-ID` to resume without losing events.",
        "operationId": "submit_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stream/{query_id}": {
      "get": {
        "tags": [
          "query"
        ],
        "summary": "Stream a query's events (SSE)",
        "operationId": "stream_stream__query_id__get",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "last-event-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream. See `GET /schema/events` for payload shapes.",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "openai"
        ],
        "summary": "Chat completion with web search and citations",
        "operationId": "chat_completions_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "system",
              "user",
              "assistant",
              "developer",
              "tool"
            ],
            "title": "Role"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "ChatMessage"
      },
      "ChatRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "default": "gleanwise-quick"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "minItems": 1,
            "title": "Messages"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          },
          "stream_options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Options"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatRequest"
      },
...
      "Focus": {
        "type": "string",
        "enum": [
          "general",
          "academic",
          "news",
          "social"
        ],
        "title": "Focus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthComponent": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint"
          },
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "name",
          "ok"
        ],
        "title": "HealthComponent"
      },
...
      "Mode": {
        "type": "string",
        "enum": [
          "quick",
          "pro",
          "deep"
        ],
        "title": "Mode"
      },
...
      "QueryAccepted": {
        "properties": {
          "query_id": {
            "type": "string",
            "title": "Query Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "stream_url": {
            "type": "string",
            "title": "Stream Url"
          }
        },
        "type": "object",
        "required": [
          "query_id",
          "thread_id",
          "stream_url"
        ],
        "title": "QueryAccepted"
      },
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Query"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode",
            "default": "quick"
          },
          "focus": {
            "$ref": "#/components/schemas/Focus",
            "default": "general"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 35
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "history": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Turn"
                },
                "type": "array",
                "maxItems": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "persist": {
            "type": "boolean",
            "title": "Persist",
            "default": true
          },
          "private": {
            "type": "boolean",
            "title": "Private",
            "default": false
          },
          "draft": {
            "type": "boolean",
            "title": "Draft",
            "default": true
          },
          "regenerate_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regenerate Message Id"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest",
        "examples": [
          {
            "query": "How do heat pumps work?"
          }
        ]
      },
...
      "SetupTestRequest": {
        "properties": {
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "llm_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Api Key"
          },
          "llm_api_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Api Base"
          },
          "searxng_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Searxng Url"
          }
        },
        "type": "object",
        "title": "SetupTestRequest"
      },
      "SetupTestResult": {
        "properties": {
          "llm": {
            "$ref": "#/components/schemas/HealthComponent"
          },
          "search": {
            "$ref": "#/components/schemas/HealthComponent"
          },
          "embeddings": {
            "$ref": "#/components/schemas/HealthComponent"
          }
        },
        "type": "object",
        "required": [
          "llm",
          "search",
          "embeddings"
        ],
        "title": "SetupTestResult"
      },
...
      "Turn": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 20000,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "Turn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}